% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tt.add.extra.cols}
\alias{tt.add.extra.cols}
\title{Add additional columns to a table.tree by applying fun on each object in obj.li}
\usage{
tt.add.extra.cols(tt, fun, obj.li = tt.obj.li(tt))
}
\description{
Add additional columns to a table.tree by applying fun on each object in obj.li
}

