% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tt.subtree}
\alias{tt.subtree}
\title{Extract subtree (as a table tree) starting at the specified row of a table tree}
\usage{
tt.subtree(tt, row = 1)
}
\description{
Extract subtree (as a table tree) starting at the specified row of a table tree
}

